requirements.txt
pkg-resources==0.0.0
pytube3==9.5.10
typing-extensions==3.7.4.1
